+++ tests/resources/CTA-North-LST-1-Released_test-telescope-model.cfg
@@ -2,12 +2,12 @@
 % TELESCOPE CONFIGURATION FILE
 % Site: North
 % ModelVersion: Released
-% TelescopeModelName: LST-1
+% TelescopeModelName: LSTN-01
 % Label: test-telescope-model
 %==================================================
 %
 #ifdef TELESCOPE
-   echo Configuration for LST-1 - TELESCOPE $(TELESCOPE)
+   echo Configuration for LSTN-01 - TELESCOPE $(TELESCOPE)
 #endif
 
 asum_clipping = 9999.0
